+++ src/components/Main/main.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from 'react-redux';
-import { Card } from "./card";
+import { FilmCard } from "./FilmCard";
 import { getFilmsIsLoading, getFilmsList } from '../../utils/redux/Films/seleсtor'; 
 import { fetchFilms } from '../../utils/redux/Films/thunks';
 
@@ -20,7 +20,7 @@ export const Main = () => {
       <h1 className="px-2 py-2 text-3xl font-bold">Афиша</h1>
       <div className="grid grid-flow-row grid-cols-3 gap-4">
         {list.map((film) => (
-          <Card key={film.id} film={film} />
+          <FilmCard key={film.id} film={film} />
         ))}
       </div>
     </main>
